requirements.txt
ccxt>=4.3.0
python-dotenv>=1.0.1
aiohttp>=3.9.5
uvloop>=0.19.0; platform_system != 'Windows'
tabulate>=0.9.0